refactor: 重构URL监听器
